today bootcamp

+++ contests/abc115/C/main.cpp
@@ -1,8 +1,9 @@
 #include <bits/stdc++.h>
-#define rep(i, n) for (int i = 0; i < n; i++)
+#define rep(i, n) for (int i = 0; i < (int)(n); i++)
 using namespace std;
 using ll = long long;
 using P = pair<int, int>;
+
 const int INF = 1001001001;
 
 int main() {
